Started changing `wavefunction.jl`
I want to find a better and nicer way to have the time dependent part of
the wave function, but I still do not know which is the best way to do
it.

Probably the best thing to do will be to put everything in a single big
function where everything is calculated, instead of having a lot of small functions.

I will try to do that in the next commit.