Reworked the time dependent part
Changed the time dependent part that goes into the integral to take only
number type arguments.
This is to make the function more general and flexible.

Finished the time dependent part of the product between the general
complex conjugato of the STA wavefunction and the time dependent part of
the ground STA state.