Single parameter test not passed
I managed to set up the code to test the result of the single paramter
and the full parameter one.

The two functions are both normalised, but they are not the same when
the time is different from 0.

I will try to calculate the corrections with the new wave function and
check the results.