Corrections not working
I tried to implement the code to evaluate the eSTA corrections but the
results are not encouraging.

I would like to try to see if I can have a numerical implementation of
the schrodinger equation to see if the Fourier transform is still a
solution of the equation in position representation.