Added sanity check to corrections
Given the number of NaN that I got when going up with the number of
particles, I decided to check if the value of the corrections contained
any `NaN`. If that was the case I would return an array of zeros.

I made use of the function `any` as well as of the function `isnan`

The scratchpad has been changed accordingly