Completed preliminary squeezing plots
overhauled the `scratchpad.jl` file so that now I can evaluate the
squeezing of the CSS state under the evolution driven by both the OAT
and the IBJJ Hamiltonian.

I compared said squeezing to the one obtained from the eSTA and STA
protcols and for some range of final times, the results are promising.