New stuff to calculate..
I need to calculate the squeezing for more particles and for the STA non
x version of the STA protocol.
This will be done by only adding a new function that use the
`control_function` instead of the `control_functionX`.
Also changed the code in such a way that both the STA and eSTA protocols
get the optimal angle before calculating the squeezing.