Found computational hiccup
It seemed like the angle for maximum sqeezing for the STA and eSTA
protocols changed sign a lot so I added an extra check to get the
minimum of the squeezing for alpha and alpha + pi / 2.

Since we agreed on only evaluating the Ueda squeezing, I am going to
remove all the parts where Jx is evaluated, to save some time.